+++ core/renderer/css/computed_css_style.h
@@ -342,6 +342,8 @@ class ComputedCSSStyle {
 
   float GetOffsetRotate() const { return offset_rotate_; }
 
+  PointerEventsType GetPointerEvents() { return pointer_events_; }
+
   static bool IsPlatformInheritableProperty(const tasm::CSSPropertyID id) {
     return GetPlatformInheritableProperty().contains(id);
   }
@@ -412,6 +414,7 @@ class ComputedCSSStyle {
   OverflowType overflow_x_{DefaultComputedStyle::DEFAULT_OVERFLOW};
   OverflowType overflow_y_{DefaultComputedStyle::DEFAULT_OVERFLOW};
   VisibilityType visibility_{DefaultComputedStyle::DEFAULT_VISIBILITY};
+  PointerEventsType pointer_events_ = PointerEventsType::kAuto;
 
   /************ css style property end ***************************/
 
@@ -548,7 +551,8 @@ class ComputedCSSStyle {
   V(XPlaceholderFontFamily)              \
   V(XPlaceholderFontSize)                \
   V(XPlaceholderFontWeight)              \
-  V(XPlaceholderFontStyle)
+  V(XPlaceholderFontStyle)               \
+  V(PointerEvents)
 
 #define GETTER_STYLE_STRING(name) lepus_value name##ToLepus();
   FOREACH_PLATFORM_PROPERTY(GETTER_STYLE_STRING)