+++ platform/darwin/ios/lynx/public/shadow_node/LynxShadowNode.h
@@ -53,6 +53,8 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (BOOL)supportInlineView;
 
+- (enum LynxPointerEventsValue)pointerEvents;
+
 @end
 
 NS_ASSUME_NONNULL_END